Moved validation code from API to ConfigManager
Reasoning here is, that the API should handle the transport part (e.g. extract payload etc.) but not process logic. This is the job of dedicated Managers - here the ConfigManager, which knows about the implementation details. Also, ConfigManager could be used directly and the validation in the API layer would be circumvented.